backend/requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.10
mongoengine==0.27.0
pymongo==4.4.0
python-dotenv==1.0.0
uvicorn==0.22.0
gunicorn==20.1.0
requests==2.31.0
beautifulsoup4==4.12.2
bilibili-api-python==1.0.0
httpx==0.28.1